refactor(login-api): create private functions, better readability
