+++ app/api/auth/_parse.ts
@@ -1,0 +1,10 @@
+/**
+ * @internal
+ * @private Only for use within /api/auth/ route.
+ */
+export function encodeRedirectParam(
+  requestSearchParams: URLSearchParams
+): string | null {
+  const redirectPath = requestSearchParams.get("redirect") || "";
+  return redirectPath ? `?redirect=${encodeURIComponent(redirectPath)}` : "";
+}
