--- lib/const.ts
@@ -1,5 +1,0 @@
-// CAS and authentication
-
-export const CAS_TEST_URL="https://secure-tst.its.yale.edu/cas";
-export const CAS_URL="https://secure-tst.its.yale.edu/cas"
-
